# Conflicts:
#	module/bdev/raid/bdev_raid.c

+++ test/make/check_so_deps.sh
@@ -4,51 +4,66 @@
 #  All rights reserved.
 #
 shopt -s extglob
-git_repo_abi="https://github.com/spdk/spdk-abi.git"
-
-function get_spdk_abi() {
-	local dest=$1
-	mkdir -p $dest
-	if [[ -d $SPDK_ABI_DIR ]]; then
-		echo "spdk-abi found at $SPDK_ABI_DIR"
-		cp -r "$SPDK_ABI_DIR"/* "$dest/"
-	else
-		# In case that someone run test manually and did not set existing
-		# spdk-abi directory via SPDK_ABI_DIR
-		echo "spdk-abi has not been found${SPDK_ABI_DIR:+ at $SPDK_ABI_DIR}, cloning"
-		git clone $git_repo_abi "$dest"
-	fi
-}
-
-function get_release_branch() {
-	tag=$(git describe --tags --abbrev=0 --exclude=LTS --exclude=*-pre $1)
-	branch="${tag:0:6}.x"
-	echo "$branch"
-}
 
 if [ "$(uname -s)" = "FreeBSD" ]; then
 	echo "Not testing for shared object dependencies on FreeBSD."
-	exit 0
-fi
-
-rootdir=$(readlink -f $(dirname $0)/../..)
-
-if [[ ! -f $1 ]]; then
-	echo "ERROR: SPDK test configuration not specified"
 	exit 1
 fi
 
-source $1
+testdir=$(readlink -f $(dirname $0))
+rootdir=$(readlink -f $testdir/../..)
+
+function usage() {
+	script_name="$(basename $0)"
+	echo "Usage: $script_name"
+	echo "    -c, --config-file     Rebuilds SPDK according to config file from autotest"
+	echo "    -a, --spdk-abi-path   Use spdk-abi from specified path, otherwise"
+	echo "                          latest version is pulled and deleted after test"
+	echo "    -h, --help            Print this help"
+	echo "Example:"
+	echo "$script_name -c ./autotest.config -a /path/to/spdk-abi"
+}
+
+# Parse input arguments #
+while getopts 'hc:a:-:' optchar; do
+	case "$optchar" in
+		-)
+			case "$OPTARG" in
+				help)
+					usage
+					exit 0
+					;;
+				config-file=*)
+					config_file="$(readlink -f ${OPTARG#*=})"
+					;;
+				spdk-abi-path=*)
+					user_abi_dir="$(readlink -f ${OPTARG#*=})"
+					;;
+				*) exit 1 ;;
+			esac
+			;;
+		h)
+			usage
+			exit 0
+			;;
+		c) config_file="$(readlink -f ${OPTARG#*=})" ;;
+		a) user_abi_dir="$(readlink -f ${OPTARG#*=})" ;;
+		*) exit 1 ;;
+	esac
+done
+
 source "$rootdir/test/common/autotest_common.sh"
 
+if [[ -e $config_file ]]; then
+	source "$config_file"
+fi
+
+source_abi_dir="${user_abi_dir:-"$testdir/abi"}"
 libdir="$rootdir/build/lib"
 libdeps_file="$rootdir/mk/spdk.lib_deps.mk"
-suppression_file="$HOME/abigail_suppressions.ini"
 
 function check_header_filenames() {
 	local dups_found=0
-
-	xtrace_disable
 
 	include_headers=$(git ls-files -- $rootdir/include/spdk $rootdir/include/spdk_internal | xargs -n 1 basename)
 	dups=
@@ -59,8 +74,6 @@ function check_header_filenames() {
 		fi
 	done
 
-	xtrace_restore
-
 	if ((dups_found == 1)); then
 		echo "Private header file(s) found with same name as public header file."
 		echo "This is not allowed since it can confuse abidiff when determining if"
@@ -70,15 +83,32 @@ function check_header_filenames() {
 	fi
 }
 
+function get_release_branch() {
+	tag=$(git describe --tags --abbrev=0 --exclude=LTS --exclude="*-pre" $1)
+	branch="${tag:0:6}.x"
+	echo "$branch"
+}
+
 function confirm_abi_deps() {
 	local processed_so=0
+	local abi_test_failed=false
 	local abidiff_output
 	local release
-	local source_abi_dir="$rootdir/test/make/abi"
+	local suppression_file="$testdir/abigail_suppressions.ini"
 
 	release=$(get_release_branch)
 
-	get_spdk_abi "$source_abi_dir"
+	if [[ ! -d $source_abi_dir ]]; then
+		mkdir -p $source_abi_dir
+		echo "spdk-abi has not been found at $source_abi_dir, cloning"
+		git clone "https://github.com/spdk/spdk-abi.git" "$source_abi_dir"
+	fi
+
+	if [[ ! -d "$source_abi_dir/$release" ]]; then
+		echo "Release (${release%.*}) does not exist in spdk-abi repository"
+		return 1
+	fi
+
 	echo "* Running ${FUNCNAME[0]} against the latest (${release%.*}) release" >&2
 
 	if ! hash abidiff; then
@@ -86,24 +116,40 @@ function confirm_abi_deps() {
 		return 1
 	fi
 
+	# Type suppression should be used for deliberate change in the structure,
+	# that do not affect the ABI compatibility.
+	# One example is utilizing a previously reserved field in the structure.
+	# Suppression file should contain at the very least the name of type
+	# and point to major SO_VER of the library it applies to.
+	#
+	# To avoid ignoring an actual changes in ABI for a particular SO_VER,
+	# the suppression should be as specific as possible to the change.
+	# has_data_member_* conditions can be used to narrow down the
+	# name of previously existing field, or specify the offset.
+	# Please refer to libabigail for details.
+	#
+	# Example format:
+	# [suppress_type]
+	#	label = Added interrupt_mode field
+	#	name = spdk_app_opts
+	#	soname_regexp = ^libspdk_event\\.so\\.12\\.*$
 	cat << EOF > ${suppression_file}
 [suppress_type]
-	name = spdk_nvme_power_state
-[suppress_type]
-	name = spdk_nvme_ctrlr_data
-[suppress_type]
-	name = spdk_nvme_cdata_oacs
-[suppress_type]
-	name = spdk_nvme_cdata_nvmf_specific
-[suppress_type]
-	name = spdk_nvme_cmd
-[suppress_type]
-	name = spdk_bs_opts
-[suppress_type]
+	label = Added interrupt_mode field
 	name = spdk_app_opts
-# To be removed, comes from nvme_internal.h
+	soname_regexp = ^libspdk_event\\.so\\.12\\.*$
+	has_data_member_regexp = ^reserved185$
+	has_data_member_inserted_between = {1480, 1488}
 [suppress_type]
-	name = spdk_nvme_qpair
+	label = Added cmbs field
+	name = spdk_nvme_cap_register
+	soname_regexp = ^libspdk_nvme\\.so\\.11\\.*$
+[suppress_type]
+	label = Added iobuf_small/large_cache_size fields
+	name = spdk_bdev_opts
+	soname_regexp = ^libspdk_bdev\\.so\\.13\\.*$
+	has_data_member_regexp = ^reserved$
+	has_data_member_inserted_between = {192, 224}
 EOF
 
 	for object in "$libdir"/libspdk_*.so; do
@@ -117,6 +163,8 @@ EOF
 
 		cmd_args=('abidiff'
 			$source_abi_dir/$release/$so_file "$libdir/$so_file"
+			'--headers-dir1' $source_abi_dir/$release/include
+			'--headers-dir2' $rootdir/include
 			'--leaf-changes-only' '--suppressions' $suppression_file)
 
 		if ! output=$("${cmd_args[@]}" --stat); then
@@ -151,7 +199,7 @@ EOF
 			if ((changed_leaf_types != 0)); then
 				if ((new_so_maj == old_so_maj)); then
 					abidiff_output=1
-					touch $fail_file
+					abi_test_failed=true
 					echo "Please update the major SO version for $so_file. A header accessible type has been modified since last release."
 				fi
 				found_abi_change=true
@@ -160,7 +208,7 @@ EOF
 			if ((removed_functions != 0)) || ((removed_vars != 0)); then
 				if ((new_so_maj == old_so_maj)); then
 					abidiff_output=1
-					touch $fail_file
+					abi_test_failed=true
 					echo "Please update the major SO version for $so_file. API functions or variables have been removed since last release."
 				fi
 				found_abi_change=true
@@ -169,7 +217,7 @@ EOF
 			if ((changed_functions != 0)) || ((changed_vars != 0)); then
 				if ((new_so_maj == old_so_maj)); then
 					abidiff_output=1
-					touch $fail_file
+					abi_test_failed=true
 					echo "Please update the major SO version for $so_file. API functions or variables have been changed since last release."
 				fi
 				found_abi_change=true
@@ -178,7 +226,7 @@ EOF
 			if ((added_functions != 0)) || ((added_vars != 0)); then
 				if ((new_so_min == old_so_min && new_so_maj == old_so_maj)) && ! $found_abi_change; then
 					abidiff_output=1
-					touch $fail_file
+					abi_test_failed=true
 					echo "Please update the minor SO version for $so_file. API functions or variables have been added since last release."
 				fi
 				found_abi_change=true
@@ -191,22 +239,22 @@ EOF
 				fi
 				if ! $found_abi_change; then
 					echo "SO name for $so_file changed without a change to abi. please revert that change."
-					touch $fail_file
+					abi_test_failed=true
 				fi
 
 				if ((new_so_maj != old_so_maj && new_so_min != 0)); then
 					echo "SO major version for $so_file was bumped. Please reset the minor version to 0."
-					touch $fail_file
+					abi_test_failed=true
 				fi
 
 				if ((new_so_min > old_so_min + 1)); then
 					echo "SO minor version for $so_file was incremented more than once. Please revert minor version to $((old_so_min + 1))."
-					touch $fail_file
+					abi_test_failed=true
 				fi
 
 				if ((new_so_maj > old_so_maj + 1)); then
 					echo "SO major version for $so_file was incremented more than once. Please revert major version to $((old_so_maj + 1))."
-					touch $fail_file
+					abi_test_failed=true
 				fi
 			fi
 
@@ -217,13 +265,18 @@ EOF
 		processed_so=$((processed_so + 1))
 	done
 	rm -f $suppression_file
+	if [[ "$processed_so" -eq 0 ]]; then
+		echo "No shared objects were processed."
+		return 1
+	fi
 	echo "Processed $processed_so objects."
-	rm -rf "$source_abi_dir"
-}
-
-function get_lib_shortname() {
-	local lib=${1##*/}
-	echo "${lib//@(libspdk_|.so)/}"
+	if [[ -z $user_abi_dir ]]; then
+		rm -rf "$source_abi_dir"
+	fi
+	if $abi_test_failed; then
+		echo "ERROR: ABI test failed"
+		exit 1
+	fi
 }
 
 function import_libs_deps_mk() {
@@ -244,6 +297,11 @@ function import_libs_deps_mk() {
 			eval "$var_mk=\${$var_mk:+\$$var_mk }$dep_mk"
 		done
 	done < "$libdeps_file"
+}
+
+function get_lib_shortname() {
+	local lib=${1##*/}
+	echo "${lib//@(libspdk_|.so)/}"
 }
 
 function confirm_deps() {
@@ -286,6 +344,9 @@ function confirm_makefile_deps() {
 	# for dependencies to avoid printing out a bunch of confusing symbols under the missing
 	# symbols section.
 	SPDK_LIBS=("$libdir/"libspdk_!(env_dpdk).so)
+	fail_file="$testdir/check_so_deps_fail"
+
+	rm -f $fail_file
 
 	declare -A IGNORED_LIBS=()
 	if grep -q 'CONFIG_RDMA?=n' $rootdir/mk/config.mk; then
@@ -297,40 +358,38 @@ function confirm_makefile_deps() {
 		for lib in "${SPDK_LIBS[@]}"; do confirm_deps "$lib" & done
 		wait
 	)
+
+	if [ -f $fail_file ]; then
+		rm -f $fail_file
+		echo "ERROR: Makefile deps test failed"
+		exit 1
+	fi
 }
 
-run_test "check_header_filenames" check_header_filenames
+if [[ -e $config_file ]]; then
+	config_params=$(get_config_params)
+	if [[ "$SPDK_TEST_OCF" -eq 1 ]]; then
+		config_params="$config_params --with-ocf=$rootdir/ocf.a"
+	fi
 
-config_params=$(get_config_params)
-if [ "$SPDK_TEST_OCF" -eq 1 ]; then
-	config_params="$config_params --with-ocf=$rootdir/ocf.a"
+	if [[ -f $rootdir/mk/config.mk ]]; then
+		$MAKE $MAKEFLAGS clean
+	fi
+
+	$rootdir/configure $config_params --with-shared
+	# By setting SPDK_NO_LIB_DEPS=1, we ensure that we won't create any link dependencies.
+	# Then we can be sure we get a valid accounting of the symbol dependencies we have.
+	SPDK_NO_LIB_DEPS=1 $MAKE $MAKEFLAGS
 fi
-
-if [[ -f $rootdir/mk/config.mk ]]; then
-	$MAKE $MAKEFLAGS clean
-fi
-
-$rootdir/configure $config_params --with-shared
-# By setting SPDK_NO_LIB_DEPS=1, we ensure that we won't create any link dependencies.
-# Then we can be sure we get a valid accounting of the symbol dependencies we have.
-SPDK_NO_LIB_DEPS=1 $MAKE $MAKEFLAGS
 
 xtrace_disable
 
-fail_file=$output_dir/check_so_deps_fail
-
-rm -f $fail_file
-
+run_test "check_header_filenames" check_header_filenames
 run_test "confirm_abi_deps" confirm_abi_deps
-
 run_test "confirm_makefile_deps" confirm_makefile_deps
 
-$MAKE $MAKEFLAGS clean
-
-if [ -f $fail_file ]; then
-	rm -f $fail_file
-	echo "shared object test failed"
-	exit 1
+if [[ -e $config_file ]]; then
+	$MAKE $MAKEFLAGS clean
 fi
 
 xtrace_restore
